Refactor ispell provider
Move IspellChecker class into ispell.cpp, leaving just provider functions in
ispell_checker.cpp. Similarly, move ispell_checker.h into ispell.h.
